python/internet-control/requirements.txt
adafruit-circuitpython-neopixel
azure-iot-device
pyaudio
rpi_ws281x